feat: implement blog retrieval and display components
