+++ client/lib/actions/blogs.ts
@@ -2,6 +2,7 @@
 
 import { AxiosError } from 'axios';
 import apiRequest from '../utils/apiRequest';
+import { Blog } from '@/types';
 
 const getTopics = async () => {
   try {
@@ -13,4 +14,14 @@ const getTopics = async () => {
   }
 };
 
-export { getTopics };
+const getUserBlogs = async () => {
+  try {
+    const res = (await apiRequest('/blogs/me')).data;
+    return { success: true, data: res.data as Blog[] };
+  } catch (error) {
+    console.log((error as AxiosError).response?.data || error);
+    return { success: false, message: 'Failed to fetch blogs' };
+  }
+};
+
+export { getTopics, getUserBlogs };
